requirements.txt
python-telegram-bot
schedule
pymongo
requests
emoji